Pam's edit
Great content! <3
Thanks for sharing. I've suggested a few edits here and there.
Should we say something about the fact that feedback can also be the start of an actual conversation, and lead to the co-construction of improvements/solutions?
I would also like to add something around empathy, putting oneself in the feedback recipient's shoes.
WDYT?